Documented code and removed code bloat
The code is now extensively documented.
The function tools.apply_accelerations_new was
the same as tools.apply_accelerations and was
removed. It will soon be replaced by a matrix-based
version. tools.initialize_coords was deprecated,
so removed as well: The user now specifies a
reasonable geometry for the animal, which is
subsequently energy minimized using the Adam
optimizer from pyTorch. The delay functionality
of tools.watch_simulation was removed.